+++ Models.cs
@@ -858,7 +858,7 @@ namespace Task
         [Serializable]
         public class Plot3D : Polyhedron
         {
-            public Plot3D(Func<double, double, double> func, int x1, int x2, int splits)
+            public Plot3D(Func<double, double, double> func, double x1, double x2, int splits)
             {
                 double step = (double)(x2 - x1) / splits;
                 var c_x = _form.pictureBox1.Image.Height / 2;
